[LLD][ELF][AArch64] Add initial .got value for R_AARCH64_GOT_PAGE_PC.
One of the differences between LLD and BFD is in the fact that BFD seems
to populate .got R_AARCH64_GOT_PAGE_PC relocation entries for which it also
adds relocations into rel[a].dyn. Such .got entries are only valid if the
ELF file is loaded at its expected load address. This allows one to write a
self-relocating ELF, like Genode's ld.lib.so.